requirements.txt
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-property==0.2.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
deprecation==2.1.0
dj-database-url==2.2.0
Django==4.2
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
minio==7.2.9
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.21.0
pydub==0.25.1
PyPDF2==3.0.1
python-decouple==3.8
python-keycloak==4.6.1
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
